\name{replaceemoji}
\alias{replaceemoji}
\title{Replace Words with Emojis}
\description{
  The user inputs a string of text, and the function finds all matching emojis for words representing emotions, replacing those words with their corresponding emojis.
}
\usage{
replaceemoji(text)
}
\arguments{
  \item{text}{A character string containing the text to be processed.}
}
\value{
  A character string with words replaced by emojis.
}
\examples{
replaceemoji("I am happy")
}
