\name{restore}
\alias{restore}
\title{Restore to Original Dataset}
\description{
  Deletes the user's custom vocabulary and restores the dataset to its original state.
}
\usage{
restore()
}
\arguments{
  \item{ }{This function does not take any arguments.}
}
\value{
  The custom vocabulary is deleted, and the original dataset is restored.
}
\examples{
restore()
}
